apirest/requirements.txt
asgiref==3.2.7
bson==0.5.8
dataclasses==0.7
Django==2.2.12
djangorestframework==3.11.0
djongo==1.3.2
dnspython==1.16.0
pkg-resources==0.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.2.4
